requirements.txt
Django==3.2.16
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
gunicorn==20.1.0
pycodestyle==2.8.0
pytz==2021.3

toml==0.10.2
whitenoise==5.3.0
django-environ==0.8.1
django-allauth==0.51.0
django-sslserver

openai~=0.25.0
django-resized == 1.0.2
Pillow == 9.3.0

#backports.zoneinfo==0.2.1
backports.zoneinfo;python_version<"3.9"
psycopg2==2.9.5
sqlparse==0.4.3

urllib3~=1.26.13
django-crispy-forms~=1.14.0
self~=2020.12.3
numpy~=1.23.5